requirements.txt
PyYAML~=5.3.1
pandas~=1.1.4
requests~=2.25.0
numpy~=1.19.4
scikit-learn~=0.23.2
pytest~=5.4.2